backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.3.2
httpx==0.27.0
python-dotenv==1.0.1
